refactor: remove extracted agent modules, delegate to data-studio-agent v0.1.1
Delete 14 agent submodules that now live in the shared library:
  chat_formatter, compact, conversation, harness, loop_runner,
  loop_runner_support, model_registry, provider_adapter,
  token_counter, tool_executor, tools, config

Add agent_adapters.rs — thin Tauri command wrappers that create
SqliteSessionStore + TauriEmitter and delegate to shared lib.

Update remaining modules:
  - executor.rs: implement ToolExecutor trait from lib
  - session_store.rs: app-specific session CRUD (no duplication)
  - capabilities/*: add parallel_ok field per lib's Capability schema
  - lib.rs: wire up Tauri commands via agent_adapters
  - Cargo.toml: data-studio-agent = { git, tag = v0.1.1 }
