+++ package.json
@@ -33,6 +33,9 @@
     "type": "git",
     "url": "https://github.com/geek-fun/dockit.git"
   },
+  "bugs": {
+    "url": "https://github.com/geek-fun/dockit/issues"
+  },
   "scripts": {
     "dev": "vite",
     "build": "vue-tsc --noEmit && vite build",
