+++ src-tauri/src/lib.rs
@@ -7,9 +7,9 @@ use tauri::AppHandle;
 pub static APP_HANDLE: OnceLock<AppHandle> = OnceLock::new();
 
 pub mod agent;
+pub mod agent_adapters;
 pub mod capabilities;
 pub mod common;
-pub mod db;
 pub mod dynamo;
 pub mod dynamo_client;
 pub mod fetch_client;
@@ -18,10 +18,6 @@ pub mod menu;
 pub mod mongo_client;
 
 use agent::executor::DocKitToolExecutor;
-use agent::loop_runner::{
-    cancel_agent_loop, compact_agent_session, confirm_tool_call, get_agent_context_usage,
-    get_tool_full_result, run_agent_loop, CancelMap, ConfirmMap,
-};
 use agent::query_history::{
     add_query_history_entry, clear_query_history, delete_query_history_entry,
     load_query_history, toggle_query_history_star,
@@ -33,11 +29,14 @@ use agent::session_store::{
     load_session_messages, migrate_session_metadata, recover_stuck_sessions,
     save_attached_source, save_confirmation_rule, update_session_meta, update_session_status,
 };
-use agent::tool_executor::ToolExecutor;
-use agent::{
-    get_all_tools, list_llm_models, run_agent_step, validate_llm_config,
+use agent_adapters::{
+    cancel_agent_loop, compact_agent_session, confirm_tool_call, get_agent_context_usage,
+    get_all_tools, get_tool_full_result, list_llm_models, run_agent_loop, run_agent_step,
+    validate_llm_config,
 };
 use capabilities::commands::{get_available_tools, invoke_capability};
+use data_studio_agent as lib;
+use data_studio_agent::storage as storage;
 use dynamo_client::{
     aws_assume_role, aws_list_profiles, aws_list_profiles_with_roles, aws_sso_get_role_credentials,
     aws_sso_list_accounts, aws_sso_list_roles, aws_sso_poll_token, aws_sso_start_device_auth,
@@ -167,8 +166,8 @@ pub fn run() {
                 .app_data_dir()
                 .map_err(|e| format!("Failed to resolve app data dir: {}", e))?;
             let db_path = app_data_dir.join("agent.sqlite");
-            let agent_db = db::open(&db_path)?;
-            db::migrate(&agent_db)?;
+            let agent_db = storage::db::open(&db_path)?;
+            storage::db::migrate(&agent_db)?;
             {
                 let conn = agent_db.0.lock().map_err(|e| e.to_string())?;
                 recover_stuck_sessions(&conn)?;
@@ -177,11 +176,11 @@ pub fn run() {
 
             use std::collections::HashMap;
             use std::sync::{Arc, Mutex};
-            let confirm_map: ConfirmMap = Arc::new(Mutex::new(HashMap::new()));
-            let cancel_map: CancelMap = Arc::new(Mutex::new(HashMap::new()));
+            let confirm_map: lib::traits::ConfirmMap = Arc::new(Mutex::new(HashMap::new()));
+            let cancel_map: lib::traits::CancelMap = Arc::new(Mutex::new(HashMap::new()));
             app.manage(confirm_map);
             app.manage(cancel_map);
-            let executor: Arc<dyn ToolExecutor> = Arc::new(DocKitToolExecutor);
+            let executor: Arc<dyn lib::ToolExecutor> = Arc::new(DocKitToolExecutor);
             app.manage(executor);
 
             use tauri::{Emitter, Listener};
